Add files via upload
This script will help you to represent the number and the origin country of Helicobacter pylori strains included in this analysis in a world map, using some R software's libraries. 

To represent the points in the map, you will need a dataset with the ubication, latitude, longitude, and number of isolates.